Readme updated to include general design of the project. Added Company as its own py file. Downloaded Netflix transcript for NLP code development before webscraping is enabled.
