+++ README.md
@@ -1,2 +1,9 @@
 # CorpEarningCallNLP
+
 Apply information extraction and sentiment analysis on public companies earning calls
+
+## work in progress
+* Tentatively thinking the classes for this project as below. 
+Drawing is done through draw.io (file in github as well)
+
+![img.png](img.png)
